pyproject.toml
[tool.poetry]
name = "ao3scraper"
version = "1.0.1"
readme = "README.md"
description = "ao3scraper is a python webscraper that scrapes AO3 for fanfiction data, stores it in a database, and highlights entries when they are updated."
license = "GPL-3.0-or-later"
authors = ["Ethan <ethanjohnleitch@gmail.com>"]
repository = "https://github.com/EthanLeitch/ao3scraper"
packages = [
    { include = "ao3scraper" },
    { include = "ao3scraper/**/*.py" },
]

[tool.poetry.scripts]
ao3scraper = "ao3scraper.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "4.11.1"
certifi = "2022.5.18.1"
chardet = "4.0.0"
click = "8.0.1"
commonmark = "0.9.1"
idna = "2.10"
Pygments = "2.12.0"
PyYAML = "6.0"
requests = "2.25.1"
rich = "12.4.1"
soupsieve = "2.3.2.post1"
urllib3 = "1.26.9"
configparser = "^5.3.0"
ao3-api = "^2.3.0"
SQLAlchemy = "^1.4.41"
marshmallow-sqlalchemy = "^0.28.1"
deepdiff = {extras = ["murmur"], version = "^5.8.1"}
dictdiffer = "^0.9.0"
pathlib = "^1.0.1"
platformdirs = "^2.5.4"
alembic = "^1.8.1"
ruamel-yaml = "^0.17.21"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
